+++ go-client-sdk/pkg/models/shared/codatdatacontractspushpushchangetypeenum.go
@@ -3,8 +3,9 @@ package shared
 type CodatDataContractsPushPushChangeTypeEnum string
 
 const (
-	CodatDataContractsPushPushChangeTypeEnumUnknown  CodatDataContractsPushPushChangeTypeEnum = "Unknown"
-	CodatDataContractsPushPushChangeTypeEnumCreated  CodatDataContractsPushPushChangeTypeEnum = "Created"
-	CodatDataContractsPushPushChangeTypeEnumModified CodatDataContractsPushPushChangeTypeEnum = "Modified"
-	CodatDataContractsPushPushChangeTypeEnumDeleted  CodatDataContractsPushPushChangeTypeEnum = "Deleted"
+	CodatDataContractsPushPushChangeTypeEnumUnknown            CodatDataContractsPushPushChangeTypeEnum = "Unknown"
+	CodatDataContractsPushPushChangeTypeEnumCreated            CodatDataContractsPushPushChangeTypeEnum = "Created"
+	CodatDataContractsPushPushChangeTypeEnumModified           CodatDataContractsPushPushChangeTypeEnum = "Modified"
+	CodatDataContractsPushPushChangeTypeEnumDeleted            CodatDataContractsPushPushChangeTypeEnum = "Deleted"
+	CodatDataContractsPushPushChangeTypeEnumAttachmentUploaded CodatDataContractsPushPushChangeTypeEnum = "AttachmentUploaded"
 )