+++ go-client-sdk/pkg/models/shared/codatpublicapimodelscompanyaddcompanymodel.go
@@ -1,6 +1,7 @@
 package shared
 
 type CodatPublicAPIModelsCompanyAddCompanyModel struct {
+	Description  *string `json:"description,omitempty"`
 	Name         string  `json:"name"`
 	PlatformType *string `json:"platformType,omitempty"`
 }