+++ python-client-sdk/src/sdk/models/shared/codat_datacontracts_push_pushchangetype_enum.py
@@ -6,4 +6,5 @@ class CodatDataContractsPushPushChangeTypeEnum(str, Enum):
     CREATED = "Created"
     MODIFIED = "Modified"
     DELETED = "Deleted"
+    ATTACHMENT_UPLOADED = "AttachmentUploaded"