+++ python-client-sdk/src/sdk/models/shared/codat_public_api_models_company_updatecompanymodel.py
@@ -1,4 +1,7 @@
 import dataclasses
+from datetime import date, datetime
+from marshmallow import fields
+import dateutil.parser
 from typing import Optional
 from dataclasses_json import dataclass_json
 from sdk import utils
@@ -6,6 +9,7 @@ from sdk import utils
 
 @dataclass_json
 @dataclasses.dataclass
-class CodatPublicAPIModelsCompanyRenameCompanyModel:
+class CodatPublicAPIModelsCompanyUpdateCompanyModel:
+    description: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.field_name('description') }})
     name: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.field_name('name') }})