+++ typescript-client-sdk/src/sdk/sdk.ts
@@ -131,8 +131,8 @@ export class SDK {
   public _securityClient: AxiosInstance;
   public _serverURL: string;
   private _language = "typescript";
-  private _sdkVersion = "1.1.0";
-  private _genVersion = "0.21.0";
+  private _sdkVersion = "1.2.2";
+  private _genVersion = "0.22.1";
 
   constructor(props: SDKProps) {
     this._serverURL = props.serverUrl ?? ServerList[0];