+++ RELEASES.md
@@ -9,3 +9,73 @@ Based on:
 - [NPM v1.1.0] https://www.npmjs.com/package/@codatio/codat-ts/v/1.1.0 - typescript-client-sdk
 - [PyPI v1.1.0] https://pypi.org/project/codatapi/1.1.0 - python-client-sdk
 - [Go v1.1.0] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.1.0 - go-client-sdk
+
+## Version 1.1.1
+### Changes
+Based on:
+- OpenAPI Doc v1 https://api.codat.io/swagger/v1/swagger.json
+- Speakeasy CLI 0.21.0 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v1.1.1] https://www.npmjs.com/package/@codatio/codat-ts/v/1.1.1 - typescript-client-sdk
+- [PyPI v1.1.1] https://pypi.org/project/codatapi/1.1.1 - python-client-sdk
+- [Go v1.1.1] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.1.1 - go-client-sdk
+
+## Version 1.1.2
+### Changes
+Based on:
+- OpenAPI Doc v1 https://api.codat.io/swagger/v1/swagger.json
+- Speakeasy CLI 0.21.0 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v1.1.2] https://www.npmjs.com/package/@codatio/codat-ts/v/1.1.2 - typescript-client-sdk
+- [PyPI v1.1.2] https://pypi.org/project/codatapi/1.1.2 - python-client-sdk
+- [Go v1.1.2] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.1.2 - go-client-sdk
+
+## Version 1.1.3
+### Changes
+Based on:
+- OpenAPI Doc v1 https://api.codat.io/swagger/v1/swagger.json
+- Speakeasy CLI 0.21.1 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v1.1.3] https://www.npmjs.com/package/@codatio/codat-ts/v/1.1.3 - typescript-client-sdk
+- [PyPI v1.1.3] https://pypi.org/project/codatapi/1.1.3 - python-client-sdk
+- [Go v1.1.3] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.1.3 - go-client-sdk
+
+## Version 1.1.4
+### Changes
+Based on:
+- OpenAPI Doc v1 https://api.codat.io/swagger/v1/swagger.json
+- Speakeasy CLI 0.21.2 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v1.1.4] https://www.npmjs.com/package/@codatio/codat-ts/v/1.1.4 - typescript-client-sdk
+- [PyPI v1.1.4] https://pypi.org/project/codatapi/1.1.4 - python-client-sdk
+- [Go v1.1.4] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.1.4 - go-client-sdk
+
+## Version 1.2.0
+### Changes
+Based on:
+- OpenAPI Doc v1 https://api.codat.io/swagger/v1/swagger.json
+- Speakeasy CLI 0.22.0 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v1.2.0] https://www.npmjs.com/package/@codatio/codat-ts/v/1.2.0 - typescript-client-sdk
+- [PyPI v1.2.0] https://pypi.org/project/codatapi/1.2.0 - python-client-sdk
+- [Go v1.2.0] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.2.0 - go-client-sdk
+
+## Version 1.2.1
+### Changes
+Based on:
+- OpenAPI Doc v1 https://api.codat.io/swagger/v1/swagger.json
+- Speakeasy CLI 0.22.0 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v1.2.1] https://www.npmjs.com/package/@codatio/codat-ts/v/1.2.1 - typescript-client-sdk
+- [PyPI v1.2.1] https://pypi.org/project/codatapi/1.2.1 - python-client-sdk
+- [Go v1.2.1] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.2.1 - go-client-sdk
+
+## Version 1.2.2
+### Changes
+Based on:
+- OpenAPI Doc v1 https://api.codat.io/swagger/v1/swagger.json
+- Speakeasy CLI 0.22.1 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v1.2.2] https://www.npmjs.com/package/@codatio/codat-ts/v/1.2.2 - typescript-client-sdk
+- [PyPI v1.2.2] https://pypi.org/project/codatapi/1.2.2 - python-client-sdk
+- [Go v1.2.2] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.2.2 - go-client-sdk